+++ eox_core/edxapp_wrapper/users.py
@@ -72,15 +72,6 @@ def get_user_signup_source():
     return backend.get_user_signup_source()
 
 
-def get_login_failures():
-    """ Gets the LoginFailures model """
-
-    backend_function = settings.EOX_CORE_USERS_BACKEND
-    backend = import_module(backend_function)
-
-    return backend.get_login_failures()
-
-
 def get_user_profile():
     """ Gets the UserProfile model """