+++ CHANGELOG.rst
@@ -12,8 +12,15 @@ Change Log
 Unreleased
 ----------
 
+Added
+~~~~~
+
 * Add proctoring test settings since this had the wrong proctoring version.
 * Adding bearer_authentication to support django-oauth2-provider and django-oauth-toolkit
+
+Removed
+~~~~~~~
+* LoginFailures andUserSignupsource admin models.
 
 [2.14.0] - 2020-09-09
 ---------------------